+++ test/conformance/corpus_data.dart
@@ -12,7 +12,7 @@ import 'dart:convert';
 String get corpusJson => utf8.decode(base64.decode(_corpusBase64));
 
 const _corpusBase64 =
-    'ewogICJ2ZXJzaW9uIjogIjEwLjAuMCIsCiAgImNhc2VzIjogWwogICAgewogICAgICAibmFtZSI6ICJzZWxlY3QgYWxsIGZy'
+    'ewogICJ2ZXJzaW9uIjogIjEwLjEuMCIsCiAgImNhc2VzIjogWwogICAgewogICAgICAibmFtZSI6ICJzZWxlY3QgYWxsIGZy'
     'b20gdGFibGUiLAogICAgICAib3BzIjogWwogICAgICAgIHsKICAgICAgICAgICJvcCI6ICJzZWxlY3RBbGwiCiAgICAgICAg'
     'fSwKICAgICAgICB7CiAgICAgICAgICAib3AiOiAiZnJvbVRhYmxlIiwKICAgICAgICAgICJ0YWJsZSI6ICJ1c2VycyIsCiAg'
     'ICAgICAgICAiYWxpYXMiOiAidSIKICAgICAgICB9CiAgICAgIF0sCiAgICAgICJleHBlY3QiOiB7CiAgICAgICAgIm1zc3Fs'